requirements.txt
# Core dependencies
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
colorama==0.4.6
rich==13.5.2
click==8.1.7

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
bandit==1.7.5
black==23.7.0
flake8==6.0.0
mypy==1.5.1
